requirements.txt
tensorflow==1.14
pillow
numpy
opencv-python
